(refactoring) AdminBookService, Repository refactoring
