+++ src/main/java/com/grepp/moodlink/app/model/data/book/BookService.java
@@ -2,30 +2,15 @@ package com.grepp.moodlink.app.model.data.book;
 
 import com.grepp.moodlink.app.model.data.book.dto.BookDto;
 import com.grepp.moodlink.app.model.data.book.entity.Book;
-import com.grepp.moodlink.app.model.data.movie.entity.Movie;
-import com.grepp.moodlink.infra.error.exceptions.CommonException;
-import com.grepp.moodlink.infra.imgbb.ImgUploadTemplate;
-import com.grepp.moodlink.infra.response.ResponseCode;
-import com.grepp.moodlink.infra.util.file.FileDto;
-import com.grepp.moodlink.infra.util.file.FileUtil;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
-
-import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.modelmapper.ModelMapper;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
-
 
 @Slf4j
 @Service
@@ -33,47 +18,6 @@ import org.springframework.web.multipart.MultipartFile;
 public class BookService {
 
     private final BookRepository bookRepository;
-    private final BookRepositoryImpl bookRepositoryImpl;
-    private final FileUtil fileUtil;
-    private final ModelMapper mapper;
-    private final ImgUploadTemplate imgUploadTemplate;
-
-    // 관리자 페이지에서 도서 추가
-    @Transactional
-    public void addBook(List<MultipartFile> thumbnail, BookDto dto) {
-
-        if(bookRepository.existsByTitleAndAuthor(dto.getTitle(),dto.getAuthor()))
-            throw new CommonException(ResponseCode.DUPLICATED_DATA);
-
-        try {
-            uploadImage(thumbnail, dto);
-
-            Book book = mapper.map(dto, Book.class);
-
-            long count = bookRepository.count();
-            book.setIsbn("B"+count);
-
-            log.info("{}",book);
-
-            bookRepository.save(book);
-        } catch (IOException e) {
-            throw new CommonException(ResponseCode.INTERNAL_SERVER_ERROR, e);
-        }
-    }
-
-    private void uploadImage(List<MultipartFile> thumbnail, BookDto dto) throws IOException {
-        if(thumbnail != null){
-            MultipartFile file =  thumbnail.getFirst();
-            String originFileName = file.getOriginalFilename();
-            if (originFileName != null && originFileName.contains(".")) {
-                String ext = originFileName.substring(originFileName.lastIndexOf("."));
-                String renameFileName = UUID.randomUUID().toString() + ext;
-
-                String thumbnailUrl = imgUploadTemplate.uploadImage(file, renameFileName);
-                dto.setImage(thumbnailUrl);
-            }
-        }
-    }
 
     public void saveMusic(List<BookDto> bookDtos) {
 
@@ -93,43 +37,6 @@ public class BookService {
         }
     }
 
-
-    public List<BookDto> findByActivated() {
-        return bookRepository.findByActivated(true)
-            .stream().map(e -> mapper.map(e, BookDto.class))
-            .toList();
-    }
-
-    public Page<BookDto> findPaged(Pageable pageable) {
-        return bookRepository.findPaged(pageable)
-            .map(e -> mapper.map(e, BookDto.class));
-    }
-
-
-    public BookDto findByIsbn(String isbn) {
-        return mapper.map(bookRepository.findByIsbn(isbn), BookDto.class);
-    }
-
-    public void updateBook(List<MultipartFile> thumbnail, BookDto dto) {
-
-        try {
-            uploadImage(thumbnail, dto);
-            // 업데이트
-            bookRepository.updateBook(dto);
-
-            log.info("{}",dto);
-
-
-        } catch (IOException e) {
-            throw new CommonException(ResponseCode.INTERNAL_SERVER_ERROR, e);
-        }
-    }
-
-    @Transactional
-    public void deleteBook(String isbn) {
-        bookRepository.findByIsbn(isbn).unActivated();
-    }
-
     public List<String> parseRecommend(String movieResult) {
         List<String> result = new ArrayList<>();
         if (movieResult == null || movieResult.isBlank()) return result;
@@ -145,10 +52,10 @@ public class BookService {
             result.add(title);
         }
         return result.stream()
-                .map(bookRepository::findIsbnByTitle)
-                .filter(Optional::isPresent)
-                .map(Optional::get)
-                .distinct()
-                .collect(Collectors.toList());
+            .map(bookRepository::findIsbnByTitle)
+            .filter(Optional::isPresent)
+            .map(Optional::get)
+            .distinct()
+            .collect(Collectors.toList());
     }
 }
